pyproject.toml
[tool.poetry]
name = "financial-analysis-backend"
version = "0.1.0"
description = "Python project to backtest/analyze stock trading strategy"
authors = ["Khoi Le <khoibaka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.41.0"
pandas = "^2.2.3"
numpy = "^2.2.0"
notebook = "^7.3.1"
Flask = "^3.1.0"
yfinance = "^0.2.50"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 